requirements.txt
langchain>=0.0.200
chromadb
streamlit
transformers
requests
python-dotenv
langchain-community
